Add error handling to Booking component
Negin & Shonaigh added the attributes to the thrown exceptions in the Booking class so the printed messages are more informative. We also added a try-catch to the loadBookings method in the DataAccessLayer to catch the exceptions thrown by the Bookings class so the component should not crash when improperly formatted data is input.

Resolves: #16